Still not working
Still struggling getting any kind of laser fire. I'm trying to learn how to use line renderer and it's not playing ball. I want to use that with Ray cast but learning from scratch and it's hard. I've also tried 3-4 other approaches and can't get it to work. I'm going to ask for advise/help in uni tomorrow and I will have it hopefully fixed within the class.